+++ src/util.cpp
@@ -10,7 +10,7 @@ void SignalBuiltinLED(int count, int delayTime)
 	}
 }
 
-int clamp(int8_t value, int8_t min, int8_t max){
+int32_t clamp(int32_t value, int32_t min, int32_t max){
 	if (value > max)
 		return max;
 	else if (value < min)
